% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_heatmap.R
\name{multimodalallostery_fitness_heatmap}
\alias{multimodalallostery_fitness_heatmap}
\title{Fitness Heatmap.}
\usage{
multimodalallostery_fitness_heatmap(input, wt_aa, title = "fitness", legend_limits = c(-1.5, 1))
}
\arguments{
\item{input}{Input file path or input data object.}

\item{wt_aa}{Value supplied for \code{wt_aa}.}

\item{title}{Value supplied for \code{title}.}

\item{legend_limits}{Value supplied for \code{legend_limits}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
