% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_energy_distance_decay_expfit_annotation.R
\name{multimodalallostery_plot_energy_distance_decay_expfit_annotation}
\alias{multimodalallostery_plot_energy_distance_decay_expfit_annotation}
\title{Plot Energy Distance Decay Expfit Annotation.}
\usage{
multimodalallostery_plot_energy_distance_decay_expfit_annotation(
  input,
  assay_sele,
  anno_file,
  x_range = c(0, 35),
  y_range = c(0, 3),
  plot_width = 4,
  plot_height = 4
)
}
\arguments{
\item{input}{Input file path or input data object.}

\item{assay_sele}{Selected assay identifier.}

\item{anno_file}{Value supplied for \code{anno_file}.}

\item{x_range}{Value supplied for \code{x_range}.}

\item{y_range}{Value supplied for \code{y_range}.}

\item{plot_width}{Value supplied for \code{plot_width}.}

\item{plot_height}{Value supplied for \code{plot_height}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
