% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_k13_vs_k19_both_promoting.R
\name{multimodalallostery_get_k13_vs_k19_both_promoting}
\alias{multimodalallostery_get_k13_vs_k19_both_promoting}
\title{Identify K13 versus K19 mutations that promote both assays}
\usage{
multimodalallostery_get_k13_vs_k19_both_promoting(
  input_k13,
  input_k19,
  anno,
  binding_sites_map,
  nbp_residues
)
}
\arguments{
\item{input_k13}{Path to the K13 input file.}

\item{input_k19}{Path to the K19 input file.}

\item{anno}{Annotation data or annotation file path.}

\item{binding_sites_map}{Named list mapping assays to binding-site residues.}

\item{nbp_residues}{Residue positions in the nucleotide-binding pocket.}
}
\value{
A data table containing mutations classified as promoting both assays.
}
\description{
Identify K13 versus K19 mutations that promote both assays
}
