% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raf1_vs_k27_anticorrelated.R
\name{multimodalallostery_get_raf1_vs_k27_anticorrelated}
\alias{multimodalallostery_get_raf1_vs_k27_anticorrelated}
\title{Identify RAF1 versus K27 anticorrelated mutations}
\usage{
multimodalallostery_get_raf1_vs_k27_anticorrelated(
  input_raf1,
  input_k27,
  anno,
  binding_sites_map,
  nbp_residues
)
}
\arguments{
\item{input_raf1}{Path to the RAF1 input file.}

\item{input_k27}{Path to the K27 input file.}

\item{anno}{Annotation data or annotation file path.}

\item{binding_sites_map}{Named list mapping assays to binding-site residues.}

\item{nbp_residues}{Residue positions in the nucleotide-binding pocket.}
}
\value{
A data table containing anticorrelated mutations.
}
\description{
Identify RAF1 versus K27 anticorrelated mutations
}
