% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapped_anticorrelated.R
\name{multimodalallostery_plot_mapped_anticorrelated}
\alias{multimodalallostery_plot_mapped_anticorrelated}
\title{Plot Mapped Anticorrelated.}
\usage{
multimodalallostery_plot_mapped_anticorrelated(
  k13_k19_data,
  raf1_k27_anticorrelated,
  nbp_residues,
  xlim = c(-1.5, 3),
  ylim = c(-1.5, 3)
)
}
\arguments{
\item{k13_k19_data}{Value supplied for \code{k13_k19_data}.}

\item{raf1_k27_anticorrelated}{Value supplied for \code{raf1_k27_anticorrelated}.}

\item{nbp_residues}{Residue positions assigned to the nucleotide-binding pocket.}

\item{xlim}{Value supplied for \code{xlim}.}

\item{ylim}{Value supplied for \code{ylim}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
