% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mapped_both_promoting_k27_k19.R
\name{multimodalallostery_plot_mapped_both_promoting_k27_k19}
\alias{multimodalallostery_plot_mapped_both_promoting_k27_k19}
\title{Plot Mapped both Promoting K27 K19.}
\usage{
multimodalallostery_plot_mapped_both_promoting_k27_k19(
  k27_k19_data,
  both_promoting_muts,
  xlim = c(-1.5, 3),
  ylim = c(-1.5, 3)
)
}
\arguments{
\item{k27_k19_data}{Value supplied for \code{k27_k19_data}.}

\item{both_promoting_muts}{Value supplied for \code{both_promoting_muts}.}

\item{xlim}{Value supplied for \code{xlim}.}

\item{ylim}{Value supplied for \code{ylim}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
