% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_dd_gb_dd_gf.R
\name{multimodalallostery_plot_scatter_dd_gb_dd_gf}
\alias{multimodalallostery_plot_scatter_dd_gb_dd_gf}
\title{Plot Scatter Dd Gb Dd Gf.}
\usage{
multimodalallostery_plot_scatter_dd_gb_dd_gf(
  ddG1,
  assay1,
  ddG2,
  assay2,
  anno,
  binder = c("K13", "K19", "RAF1"),
  colour_scheme,
  xlim = c(-1, 3.3),
  ylim = c(-1.7, 3)
)
}
\arguments{
\item{ddG1}{Value supplied for \code{ddG1}.}

\item{assay1}{Value supplied for \code{assay1}.}

\item{ddG2}{Value supplied for \code{ddG2}.}

\item{assay2}{Value supplied for \code{assay2}.}

\item{anno}{Annotation data or annotation file path used by the analysis.}

\item{binder}{Value supplied for \code{binder}.}

\item{colour_scheme}{Value supplied for \code{colour_scheme}.}

\item{xlim}{Value supplied for \code{xlim}.}

\item{ylim}{Value supplied for \code{ylim}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
