% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sequence_annotation.R
\name{multimodalallostery_plot_sequence_annotation}
\alias{multimodalallostery_plot_sequence_annotation}
\title{Plot Sequence Annotation.}
\usage{
multimodalallostery_plot_sequence_annotation(
  prepared_sequence,
  gtp_pocket,
  functional_loops,
  core_residues,
  beta_sheets,
  alpha_helices
)
}
\arguments{
\item{prepared_sequence}{Value supplied for \code{prepared_sequence}.}

\item{gtp_pocket}{Value supplied for \code{gtp_pocket}.}

\item{functional_loops}{Value supplied for \code{functional_loops}.}

\item{core_residues}{Value supplied for \code{core_residues}.}

\item{beta_sheets}{Value supplied for \code{beta_sheets}.}

\item{alpha_helices}{Value supplied for \code{alpha_helices}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
