% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_site_scatter.R
\name{multimodalallostery_plot_site_scatter}
\alias{multimodalallostery_plot_site_scatter}
\title{Plot Site Scatter.}
\usage{
multimodalallostery_plot_site_scatter(
  full_analysis_result,
  target_position,
  point_size = 4,
  alpha = 0.8,
  base_size = 14,
  show_labels = TRUE,
  label_all = TRUE,
  xlim = NULL,
  ylim = NULL,
  show_WT = TRUE,
  WT_point_size = 3,
  WT_color = "black",
  WT_shape = 19,
  WT_label = "WT"
)
}
\arguments{
\item{full_analysis_result}{Value supplied for \code{full_analysis_result}.}

\item{target_position}{Value supplied for \code{target_position}.}

\item{point_size}{Value supplied for \code{point_size}.}

\item{alpha}{Value supplied for \code{alpha}.}

\item{base_size}{Value supplied for \code{base_size}.}

\item{show_labels}{Value supplied for \code{show_labels}.}

\item{label_all}{Value supplied for \code{label_all}.}

\item{xlim}{Value supplied for \code{xlim}.}

\item{ylim}{Value supplied for \code{ylim}.}

\item{show_WT}{Value supplied for \code{show_WT}.}

\item{WT_point_size}{Value supplied for \code{WT_point_size}.}

\item{WT_color}{Value supplied for \code{WT_color}.}

\item{WT_shape}{Value supplied for \code{WT_shape}.}

\item{WT_label}{Value supplied for \code{WT_label}.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Reusable plotting logic consolidated from the selected Figure/Panel implementations.
}
