% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_beta_sheet_ddg.R
\name{multimodalallostery_prepare_beta_sheet_ddg}
\alias{multimodalallostery_prepare_beta_sheet_ddg}
\title{Prepare Beta Sheet Ddg.}
\usage{
multimodalallostery_prepare_beta_sheet_ddg(
  ddg,
  annotation,
  beta_sheet_ranges,
  sheet_levels = c("b2", "b3", "b1", "b4", "b5", "b6")
)
}
\arguments{
\item{ddg}{Value supplied for \code{ddg}.}

\item{annotation}{Annotation data frame.}

\item{beta_sheet_ranges}{Value supplied for \code{beta_sheet_ranges}.}

\item{sheet_levels}{Value supplied for \code{sheet_levels}.}
}
\value{
A prepared data object used by downstream analysis.
}
\description{
Reusable data preparation logic consolidated from the selected Figure/Panel implementations.
}
